update other requirements
